requirements.txt
anthropic>=0.34.0
flask>=3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
google-generativeai>=0.3.0





